♻️ rename project projectors
