--- marketplace-domain/src/projectors/member/mod.rs
@@ -1,6 +1,0 @@
-#[allow(clippy::module_inception)]
-mod member;
-pub use member::*;
-
-#[cfg(test)]
-mod test;
