+++ marketplace-domain/src/projectors/mod.rs
@@ -1,8 +1,8 @@
 mod contributor_with_github_data;
 pub use contributor_with_github_data::ContributorWithGithubData;
 
-mod lead_projector;
-pub use lead_projector::LeadContributorProjector;
+mod project_lead_contributors;
+pub use project_lead_contributors::LeadContributorProjector;
 
-mod member;
-pub use member::MemberProjector;
+mod project_members;
+pub use project_members::MemberProjector;
